Разбивает reducer на отдельные
