+++ src/lib/is-object/is-object.js
@@ -1,0 +1,8 @@
+/**
+ * Проверяет является ли значение объектом Object
+ * @param {*} value
+ * @return {boolean}
+ */
+export const isObject = (value) => {
+  return value && typeof value === `object` && value.constructor === Object;
+};
