+++ src/reducer/index.js
@@ -1,0 +1,9 @@
+import {combineReducers} from 'redux';
+import {reducer as data} from './data/data';
+import {reducer as user} from './user/user';
+
+
+export default combineReducers({
+  DATA: data,
+  USER: user,
+});
